[REFACTOR] 기존 HomeActivity 대체
